+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvIdCoordinateSource.java
@@ -52,8 +52,7 @@ public class CsvIdCoordinateSource extends CsvDataSource implements IdCoordinate
     return buildStreamWithFieldsToAttributesMap()
         .map(fieldToValues -> new SimpleFactoryData(fieldToValues, Pair.class))
         .map(factory::get)
-        .filter(Optional::isPresent)
-        .map(Optional::get)
+        .flatMap(Optional::stream)
         .collect(Collectors.toMap(Pair::getKey, Pair::getValue));
   }